+++ lib/view_component/base.rb
@@ -6,6 +6,7 @@ require "view_component/collection"
 require "view_component/compile_cache"
 require "view_component/compiler"
 require "view_component/config"
+require "view_component/errors"
 require "view_component/preview"
 require "view_component/slotable"
 require "view_component/translatable"
@@ -27,8 +28,6 @@ module ViewComponent
       # unless you're building a `ViewComponent::Config` elsewhere.
       attr_writer :config
     end
-
-    class ViewContextCalledBeforeRenderError < StandardError; end
 
     include ViewComponent::Slotable
     include ViewComponent::Translatable
@@ -175,16 +174,7 @@ module ViewComponent
     #
     # @return [ActionController::Base]
     def controller
-      if view_context.nil?
-        raise(
-          ViewContextCalledBeforeRenderError,
-          "`#controller` can't be used during initialization, as it depends " \
-          "on the view context that only exists once a ViewComponent is passed to " \
-          "the Rails render pipeline.\n\n" \
-          "It's sometimes possible to fix this issue by moving code dependent on " \
-          "`#controller` to a `#before_render` method: https://viewcomponent.org/api.html#before_render--void."
-        )
-      end
+      raise(ControllerCalledBeforeRenderError) if view_context.nil?
 
       @__vc_controller ||= view_context.controller
     end
@@ -196,7 +186,7 @@ module ViewComponent
     def helpers
       if view_context.nil?
         raise(
-          ViewContextCalledBeforeRenderError,
+          ViewComponent::ViewContextCalledBeforeRenderError,
           "`#helpers` can't be used during initialization as it depends " \
           "on the view context that only exists once a ViewComponent is passed to " \
           "the Rails render pipeline.\n\n" \